[ADD/UPT][#11] Added search bar.
