+++ vue-front/src/App.vue
@@ -1,74 +1,50 @@
 <template>
   <div class="app justify-center">
-    <v-row no-gutters class="justify-center">
-      <v-col
-        v-bind:key="item"
-        v-for="item in plotItems"
-        class="plotly-chart center-content"
+    <!-- <v-responsive class="mx-auto" max-width="344"> -->
+    <v-responsive id="search-bar" max-width="700">
+      <v-text-field
+        v-model="username"
+        label="Search for a GitHub username"
+        hint="Write a GitHub username and press enter to visualize their account information"
+        placeholder="ArielMAJ"
+        clearable
+        type="text"
+        variant="solo-filled"
+        :rules="[rules.required]"
+        @keyup.enter="search"
       >
-        <h2 class="margin-up" v-if="item.chartData !== null">
-          {{ item.chartData[0].labels[0] }}
-        </h2>
-        <Loading v-if="item.chartData === null" />
-        <PlotlyChart
-          v-else
-          :chartData="item.chartData"
-          :divId="item.information"
-        />
-      </v-col>
-    </v-row>
+      </v-text-field>
+    </v-responsive>
+    <!-- <v-btn
+      class="white--text"
+      color="primary"
+      @click="pieChartModalOpen = !pieChartModalOpen"
+    >Open Repository Visualization
+    </v-btn> -->
+    <PieChartModal v-if="pieChartModalOpen" :username="this.username" />
   </div>
 </template>
 <script>
-import Loading from "./components/Loading.vue";
-import PlotlyChart from "./components/PlotlyChart.vue";
+import PieChartModal from "./modals/PieChartModal.vue";
 
 export default {
   name: "app",
   components: {
-    PlotlyChart,
-    Loading,
+    PieChartModal,
   },
   data: function () {
     return {
-      plotItems: [
-        { information: "allow_forking", chartData: null },
-        { information: "archived", chartData: null },
-        { information: "disabled", chartData: null },
-        { information: "fork", chartData: null },
-        { information: "has_discussions", chartData: null },
-        { information: "has_downloads", chartData: null },
-        { information: "has_issues", chartData: null },
-        { information: "has_pages", chartData: null },
-        { information: "has_projects", chartData: null },
-        { information: "has_wiki", chartData: null },
-        { information: "is_template", chartData: null },
-        { information: "private", chartData: null },
-        { information: "web_commit_signoff_required", chartData: null },
-      ],
+      pieChartModalOpen: false,
+      rules: {
+        required: (value) => !!value || "Field is required",
+      },
+      username: "",
     };
   },
-  async mounted() {
-    this.plotItems.forEach((item) => {
-      this.fetchData("a", item);
-    });
-  },
   methods: {
-    async fetchData(username, item) {
-      try {
-        console.log(
-          "Fetching data...",
-          process.env.VUE_APP_BACKEND_ROOT_ENDPOINT +
-            `${username}/${item.information}`
-        );
-        const resp = await fetch(
-          process.env.VUE_APP_BACKEND_ROOT_ENDPOINT +
-            `${username}/${item.information}`
-        );
-        item.chartData = await resp.json();
-        console.log(item);
-      } catch (error) {
-        console.error("Error fetching data:", error);
+    search() {
+      if (this.username !== "") {
+        this.pieChartModalOpen = true;
       }
     },
   },
@@ -86,11 +62,13 @@ export default {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
-  color: #2c3e50;
+  color: rgb(4, 162, 4);
   background-color: black;
   min-height: 100vh;
   min-width: 100vw;
   max-width: 100vw;
+  margin: 0;
+  position: absolute;
 }
 
 body {
@@ -100,31 +78,16 @@ body {
   max-width: 100vw;
 }
 
-.plotly-chart {
-  margin: 20px;
-  border: 1px solid #ccc;
-  border-radius: 50px;
-  background-color: transparent;
-  height: 325px;
-  min-width: 325px;
-  max-width: 40%;
-}
-
-.margin-up {
-  margin-top: 20px;
-}
-.center-content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-}
-
 .justify-center {
   justify-content: center;
 }
 
 ::-webkit-scrollbar {
   display: none;
+}
+
+#search-bar {
+  margin: auto;
+  margin-top: 20px;
 }
 </style>
